bot/requirement.txt
autopep8
pylint

python-telegram-bot
requests
python-dotenv
paramiko

psutil
